2017113 kenura

+++ back-end/E-Learning_Platform/E-Learning_Platform/src/main/java/CS2001/Group47/ELearning_Platform/repository/StudentRepository.java
@@ -20,6 +20,6 @@ public interface StudentRepository extends JpaRepository<Student, Integer> {
 
 	Student findByResetPasswordToken(String token);
 
-	Student findByStudentId(Long userId);
+	Student findByStudentId(Integer userId);
 
 }
